feat(rule-finder): Omit deprecated rules
Filter out deprecated rules in `_getAllAvailableRules()` and
`_getPluginRules()`.  With this change, the lists of rules
displayed with the `-a` and `-u` flags will not include
deprecated rules.

Breaking: The output is now different from what it was, as
deprecated rules are now omitted by default.

Breaking: The `getRuleFinder()` API has changed.  Previously, it
took an optional filename and an optional boolean, `noCore`.
Now, it takes the optional filename and an options object:

* `omitCore`: Exclude core rules from the output
* `includeDeprecated`: Include deprecated rules in the output

Both options default to `false` if they are not provided.

Per a review request, I've added a flag to allow deprecated rules
to be included.  Adding `--include deprecated` or `-i deprecated`
will include deprecated rules in the output; they're not included
by default.

To make the implementation of this flag cleaner, I refactored
`getRuleFinder()` to take an options object instead of a boolean
as the second argument.  Adding a second boolean
(`includeDeprecated`) made it a difficult interface to use.

Note that `eslint.linter.getRules()` returns an ES6 Map, which
cannot be directly filtered.  Also, its `keys()` function returns
an iterable, so we have to spread that into an array before we
can filter it.  That's why the code is written the way it is.

To make this feature have a cleaner implementation, I also
performed the following refactorings:

* Don't pass plugin rules to `_getAllAvailableRules`.  Plugin
  rules are now merged with core rules directly in `RuleFinder`
  instead.

* Rename `_getAllAvailableRules` -> `_getCoreRules` to better
  reflect its purpose.

I also:

* Removed the `nc` option; it isn't documented and isn't being
  used.

* Split some of the `find` tests into two `it` blocks.  I found
  that calling `proxyquire` twice in the same test wasn't doing
  anything, so was potentially hiding failures.

* Renamed some `rule-finder` tests to result in more consistent
  output.